There is no need to separate an interface API
